+++ src/migrations/M231211093758Indexes.php
@@ -2,7 +2,9 @@
 
 namespace davidhirtz\yii2\media\migrations;
 
+use davidhirtz\yii2\media\models\File;
 use davidhirtz\yii2\media\models\Folder;
+use davidhirtz\yii2\media\models\Transformation;
 use davidhirtz\yii2\skeleton\db\traits\MigrationTrait;
 use yii\db\Migration;
 
@@ -16,11 +18,15 @@ class M231211093758Indexes extends Migration
 
     public function safeUp(): void
     {
+        $this->dropForeignKey('folder_parent_id_ibfk', Folder::tableName());
+
         $this->dropColumn(Folder::tableName(), 'parent_id');
         $this->dropColumn(Folder::tableName(), 'lft');
         $this->dropColumn(Folder::tableName(), 'rgt');
 
         $this->createIndex('path', Folder::tableName(), 'path', true);
+        $this->createIndex('basename', File::tableName(), ['basename', 'folder_id', 'extension'], true);
+        $this->createIndex('name', Transformation::tableName(), ['name', 'file_id', 'extension'], true);
 
         parent::safeUp();
     }
@@ -28,10 +34,20 @@ class M231211093758Indexes extends Migration
     public function safeDown(): void
     {
         $this->dropIndex('path', Folder::tableName());
+        $this->dropIndex('basename', File::tableName());
+        $this->dropIndex('name', Transformation::tableName());
 
-        $this->addColumn(Folder::tableName(), 'parent_id', $this->integer()->null()->after('type'));
-        $this->addColumn(Folder::tableName(), 'lft', $this->integer()->null()->after('parent_id'));
-        $this->addColumn(Folder::tableName(), 'rgt', $this->integer()->null()->after('lft'));
+        $this->addColumn(Folder::tableName(), 'parent_id', $this->integer()->unsigned()->null()->after('type'));
+        $this->addColumn(Folder::tableName(), 'lft', $this->integer()->unsigned()->null()->after('parent_id'));
+        $this->addColumn(Folder::tableName(), 'rgt', $this->integer()->unsigned()->null()->after('lft'));
+
+        $this->addForeignKey('folder_parent_id_ibfk',
+            Folder::tableName(),
+            'parent_id',
+            Folder::tableName(),
+            'id',
+            'SET NULL'
+        );
 
         parent::safeDown();
     }